requirements.txt
langchain==0.1.14
langchain-core==0.1.37
langchain-google-genai==0.0.6
python-dotenv==1.0.1
pandas==2.2.1
pillow==11.2.1
pydantic==2.11.3
unstructured==0.17.2
matplotlib==3.10.1
langchain-openai
langchain-huggingface
